requirements.txt
PySide6
Pillow
rawpy
pillow-heif
